+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 3.3.0
 
 -   Update `openai` dependency
+-   Specify that old configuration options are for Codex only
+-   Change namespace for configuration options from `chatgptHelper` to `chatgpt-helper` to be in line with rest of extension
 
 ## 3.2.1